requirements.txt
pip==23.3.2
streamlit==1.29.0
google-generativeai==0.3.2
python-dotenv==1.0.0
langchain==0.1.0
PyPDF2==1.26.0
chromadb==0.4.22
faiss-cpu==1.7.4
gtts==2.2.3